refactor: PathValidator 삭제
