+++ src/main/java/domain/Board.java
@@ -5,18 +5,17 @@ import domain.piece.Piece;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class Board {
 
     private static final String IMPOSSIBLE_MOVE_ERROR_MESSAGE = "이동할 수 없는 위치입니다.";
     private static final String WHITE_TURN_ERROR_MESSAGE = "흰 진영 차례입니다.";
     private static final String BLACK_TURN_ERROR_MESSAGE = "검은 진영 차례입니다.";
-    private final PathValidator pathValidator;
     private final Map<Location, Piece> board;
 
 
-    public Board(final PathValidator pathValidator, final Map<Location, Piece> board) {
-        this.pathValidator = pathValidator;
+    public Board(final Map<Location, Piece> board) {
         this.board = board;
     }
 
@@ -60,15 +59,9 @@ public class Board {
         final Section start = Section.of(startLocation, startPiece);
         final Section end = Section.of(endLocation, endPiece);
         final List<Piece> path = getPiecesInPath(start, end);
-        if (pathValidator.isValid(start, end, path)) {
-            return;
+        if (isBlocked(path)) {
+            throw new IllegalArgumentException(IMPOSSIBLE_MOVE_ERROR_MESSAGE);
         }
-        throw new IllegalArgumentException(IMPOSSIBLE_MOVE_ERROR_MESSAGE);
-    }
-
-    private void convert(final Location start, final Location end) {
-        board.replace(end, board.get(start));
-        board.replace(start, EmptyPiece.make());
     }
 
     private List<Piece> getPiecesInPath(final Section start, final Section end) {
@@ -77,6 +70,17 @@ public class Board {
         return paths.stream()
             .map(this::findPiece)
             .collect(Collectors.toList());
+    }
+
+    private boolean isBlocked(final List<Piece> path) {
+        return IntStream.range(0, path.size() - 1)
+            .mapToObj(path::get)
+            .anyMatch(Piece::isNotEmpty);
+    }
+
+    private void convert(final Location start, final Location end) {
+        board.replace(end, board.get(start));
+        board.replace(start, EmptyPiece.make());
     }
 
     public Piece findPiece(final Location location) {
