+++ src/test/java/domain/BoardTest.java
@@ -22,7 +22,7 @@ class BoardTest {
     public void testCreate() {
         //given
         //when
-        final Board board = new Board(new PathValidator(), new HashMap<>());
+        final Board board = new Board(new HashMap<>());
         board.initialize();
 
         //then
@@ -41,7 +41,7 @@ class BoardTest {
         @DisplayName("흰 진영 순서일 때 검은 진영을 움직인다.")
         @Test
         public void testMoveFailBlack() {
-            final Board board = new Board(new PathValidator(), new HashMap<>());
+            final Board board = new Board(new HashMap<>());
             board.initialize();
 
             assertThatThrownBy(
@@ -52,7 +52,7 @@ class BoardTest {
         @DisplayName("검은 진영 순서일 때 흰 진영을 움직인다.")
         @Test
         public void testMoveFailWhite() {
-            final Board board = new Board(new PathValidator(), new HashMap<>());
+            final Board board = new Board(new HashMap<>());
             board.initialize();
 
             assertThatThrownBy(
@@ -69,7 +69,7 @@ class BoardTest {
         @DisplayName("흰생 진영의 순서일 때 흰색 진영의 돌을 움직인다.")
         public void testMoveWhite() {
             //given
-            final Board board = new Board(new PathValidator(), new HashMap<>());
+            final Board board = new Board(new HashMap<>());
             board.initialize();
 
             //when
@@ -84,7 +84,7 @@ class BoardTest {
         @DisplayName("검은색 진영의 순서일 때 검은색 진영의 돌을 움직인다.")
         public void testMoveBlack() {
             //given
-            final Board board = new Board(new PathValidator(), new HashMap<>());
+            final Board board = new Board(new HashMap<>());
             board.initialize();
 
             //when
